Abstract functions and constants to the new data and logic folders
