+++ src/components/AboutLists.js
@@ -1,49 +1,13 @@
 import AboutList from './AboutList';
+import experiences from '../data/experiences';
 
-const AboutLists = () => {
-  const experiences = [
-    {
-      id: 1,
-      name: 'Languages',
-      items: [
-        { id: 1, skill: 'JavaScript (ES6)' },
-        { id: 2, skill: 'Ruby' },
-        { id: 3, skill: 'HTML5' },
-        { id: 4, skill: 'CSS3' },
-      ],
-    },
-    {
-      id: 2,
-      name: 'FrameWorks',
-      items: [
-        { id: 1, skill: 'Ruby on Rails' },
-        { id: 2, skill: 'Bootstrap' },
-        { id: 3, skill: 'React' },
-        { id: 4, skill: 'Rspec' },
-        { id: 5, skill: 'Jest' },
-      ],
-    },
-    {
-      id: 3,
-      name: 'Skills',
-      items: [
-        { id: 1, skill: 'Remote Collaboration' },
-        { id: 2, skill: 'Pair-Programming' },
-        { id: 3, skill: 'Git' },
-        { id: 4, skill: 'GitHub' },
-        { id: 5, skill: 'Terminal' },
-      ],
-    },
-  ];
-
-  return (
-    <div className="about-lists">
-      {experiences.map((experience) => (
-        <AboutList key={experience.id} experience={experience} />
-      ))}
-      <div className="about-deco" />
-    </div>
-  );
-};
+const AboutLists = () => (
+  <div className="about-lists">
+    {experiences.map((experience) => (
+      <AboutList key={experience.id} experience={experience} />
+    ))}
+    <div className="about-deco" />
+  </div>
+);
 
 export default AboutLists;
