+++ src/data/navLinks.js
@@ -1,0 +1,24 @@
+const navLinks = [
+  {
+    id: 'hello',
+    path: '#hello',
+    text: 'Hello',
+  },
+  {
+    id: 'portfolio',
+    path: '#portfolio',
+    text: 'Projects',
+  },
+  {
+    id: 'about',
+    path: '#about',
+    text: 'About',
+  },
+  {
+    id: 'contact',
+    path: '#contact',
+    text: 'Contact',
+  },
+];
+
+export default navLinks;
