+++ src/logic/isMedScrn.js
@@ -1,0 +1,8 @@
+const isMedScrn = () => {
+  if (window.innerWidth < 769) {
+    return false;
+  }
+  return true;
+};
+
+export default isMedScrn;
