+++ src/style.scss
@@ -972,8 +972,8 @@ footer {
     max-width: 775px;
   }
 
-  .about {
-    justify-content: space-evenly;
+  .about-title {
+    padding-top: 45px;
   }
 
   form {
